pyproject.toml
[project]
name = "carbon-check-field"
version = "1.0.0"
description = "Carbon credit estimation for agricultural fields using ML crop classification"
requires-python = ">=3.9,<3.14"
# Note: This project uses separate requirements.txt files for each GCP service
# See individual requirements.txt files:
# - backend/requirements.txt (for Cloud Run)
# - ml_pipeline/orchestrator/requirements.txt (for Cloud Run orchestrator)
# - ml_pipeline/trainer/requirements.txt (for Vertex AI training)
# This pyproject.toml is for LOCAL DEVELOPMENT ONLY
dependencies = [
    # Core dependencies (required for all components)
    "PyYAML>=6.0.1",
    "numpy>=1.24.0,<2.0.0",  # Must be <2.0.0 for sklearn-cpu.1-3 container compatibility
    "pandas>=2.0.0",
    
    # Google Cloud SDK (core)
    "google-cloud-bigquery>=3.13.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-aiplatform>=1.38.0",
    "earthengine-api>=0.1.384",
    "google-auth>=2.27.0",
    "db-dtypes>=1.2.0",
]

[project.optional-dependencies]
# Backend API dependencies (matches backend/requirements.txt)
backend = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "pydantic==2.5.3",
    "shapely==2.0.1",
    "gunicorn==21.2.0",
]

# ML Training dependencies (matches ml_pipeline/trainer/requirements.txt)
trainer = [
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tensorboard>=2.14.0",
    "torch>=2.0.0",  # For SummaryWriter
    "pillow>=10.0.0",
    "gcsfs>=2023.6.0",
]

# Orchestrator dependencies (matches ml_pipeline/orchestrator/requirements.txt)
orchestrator = [
    "flask>=2.3.0",
    "gunicorn>=21.2.0",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

# All dependencies (for local development/testing)
all = [
    "carbon-check-field[backend,trainer,orchestrator]",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ml_pipeline", "backend"]
